// src/store/complexData/type.ts
export interface complexDataType {
  userRemark: any[]
  newestRemark: any[]
  showloadings: boolean
  useralbumslist: any
  useralbumsMusics: any[]
  Hotspots: any[]
  userAlbumItem: any
  artistDetail: any
  SimilarArtist: any[]
  artistmvs: any[]
  MVurl: any
  mvdetail: any
  similarmv: any[]
  remarknumber: number
  ThreadId: any
}